Refactored the PackageBuilder logic into a Provider, to reduce complexity of the component itself.
