+++ ui.frontend/src/utility/hooks/useDebounce.ts
@@ -1,12 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
-
-export function useRenderCount(): number {
-  const count = useRef(0);
-
-  count.current++;
-
-  return count.current;
-}
+import { useEffect, useState } from 'react';
 
 export function useDebounce<T>(value: T, delay: number): T {
   const [debouncedValue, setDebouncedValue] = useState(value);
